prepare windows support for cni server
