+++ pkg/daemon/config.go
@@ -13,7 +13,6 @@ import (
 
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/pflag"
-	"github.com/vishvananda/netlink"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
@@ -151,7 +150,6 @@ func (config *Configuration) initNicConfig(nicBridgeMappings map[string]string) 
 	}
 
 	isDPDKNode := node.GetLabels()[util.OvsDpTypeLabel] == "userspace"
-
 	if isDPDKNode {
 		config.Iface = config.DPDKTunnelIface
 	}
@@ -263,33 +261,6 @@ func (config *Configuration) initKubeClient() error {
 	}
 	config.KubeClient = kubeClient
 	return nil
-}
-
-func getIfaceOwnPodIP(podIP string) (*net.Interface, error) {
-	links, err := netlink.LinkList()
-	if err != nil {
-		return nil, err
-	}
-
-	for _, link := range links {
-		addrs, err := netlink.AddrList(link, netlink.FAMILY_ALL)
-		if err != nil {
-			return nil, fmt.Errorf("failed to get a list of IP addresses %v", err)
-		}
-		for _, addr := range addrs {
-			if addr.IPNet.Contains(net.ParseIP(podIP)) && addr.IP.String() == podIP {
-				return &net.Interface{
-					Index:        link.Attrs().Index,
-					MTU:          link.Attrs().MTU,
-					Name:         link.Attrs().Name,
-					HardwareAddr: link.Attrs().HardwareAddr,
-					Flags:        link.Attrs().Flags,
-				}, nil
-			}
-		}
-	}
-
-	return nil, errors.New("unable to find podIP interface")
 }
 
 func setEncapIP(ip string) error {
