+++ pkg/daemon/handler.go
@@ -4,19 +4,15 @@ import (
 	"context"
 	"fmt"
 	"net/http"
-	"os"
-	"path"
 	"strconv"
 	"strings"
 	"time"
 
 	"github.com/emicklei/go-restful/v3"
-	"golang.org/x/sys/unix"
 	v1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/klog/v2"
 
@@ -230,7 +226,7 @@ func (csh cniServerHandler) handleAdd(req *restful.Request, resp *restful.Respon
 		if nicType == util.InternalType {
 			podNicName, err = csh.configureNicWithInternalPort(podRequest.PodName, podRequest.PodNamespace, podRequest.Provider, podRequest.NetNs, podRequest.ContainerID, ifName, macAddr, mtu, ipAddr, gw, isDefaultRoute, podRequest.Routes, ingress, egress, priority, podRequest.DeviceID, nicType, gatewayCheckMode)
 		} else if nicType == util.DpdkType {
-			err = csh.configureDpdkNic(podRequest.PodName, podRequest.PodNamespace, podRequest.Provider, podRequest.NetNs, podRequest.ContainerID, ifName, macAddr, mtu, ipAddr, gw, ingress, egress, priority, path.Join("/var", getShortSharedDir(pod.UID, podRequest.VhostUserSocketVolumeName)), podRequest.VhostUserSocketName)
+			err = csh.configureDpdkNic(podRequest.PodName, podRequest.PodNamespace, podRequest.Provider, podRequest.NetNs, podRequest.ContainerID, ifName, macAddr, mtu, ipAddr, gw, ingress, egress, priority, getShortSharedDir(pod.UID, podRequest.VhostUserSocketVolumeName), podRequest.VhostUserSocketName)
 		} else {
 			podNicName = ifName
 			err = csh.configureNic(podRequest.PodName, podRequest.PodNamespace, podRequest.Provider, podRequest.NetNs, podRequest.ContainerID, podRequest.VfDriver, ifName, macAddr, mtu, ipAddr, gw, isDefaultRoute, podRequest.Routes, ingress, egress, priority, podRequest.DeviceID, nicType, gatewayCheckMode)
@@ -263,66 +259,6 @@ func (csh cniServerHandler) handleAdd(req *restful.Request, resp *restful.Respon
 	if err := resp.WriteHeaderAndEntity(http.StatusOK, request.CniResponse{Protocol: util.CheckProtocol(cidr), IpAddress: ip, MacAddress: macAddr, CIDR: cidr, Gateway: gw, PodNicName: podNicName}); err != nil {
 		klog.Errorf("failed to write response, %v", err)
 	}
-}
-
-func createShortSharedDir(pod *v1.Pod, volumeName string) (err error) {
-	var volume *v1.Volume
-	for index, v := range pod.Spec.Volumes {
-		if v.Name == volumeName {
-			volume = &pod.Spec.Volumes[index]
-			break
-		}
-	}
-	if volume == nil {
-		return fmt.Errorf("can not found volume %s in pod %s", volumeName, pod.Name)
-	}
-	if volume.EmptyDir == nil {
-		return fmt.Errorf("volume %s is not empty dir", volume.Name)
-	}
-	originSharedDir := fmt.Sprintf("/var/lib/kubelet/pods/%s/volumes/kubernetes.io~empty-dir/%s", pod.UID, volumeName)
-	newSharedDir := getShortSharedDir(pod.UID, volumeName)
-	if _, err = os.Stat(newSharedDir); os.IsNotExist(err) {
-		err = os.MkdirAll(newSharedDir, 0750)
-		if err != nil {
-			return fmt.Errorf("createSharedDir: Failed to create dir (%s): %v", newSharedDir, err)
-		}
-
-		if strings.Contains(newSharedDir, util.DefaultHostVhostuserBaseDir) {
-			klog.Infof("createSharedDir: Mount from %s to %s", originSharedDir, newSharedDir)
-			err = unix.Mount(originSharedDir, newSharedDir, "", unix.MS_BIND, "")
-			if err != nil {
-				return fmt.Errorf("createSharedDir: Failed to bind mount: %s", err)
-			}
-		}
-		return nil
-
-	}
-	return err
-
-}
-
-func removeShortSharedDir(pod *v1.Pod, volumeName string) (err error) {
-	sharedDir := getShortSharedDir(pod.UID, volumeName)
-	if _, err = os.Stat(sharedDir); os.IsNotExist(err) {
-		klog.Errorf("shared directory %s does not exist to unmount, %s", sharedDir, err)
-		return nil
-	}
-	err = unix.Unmount(sharedDir, 0)
-	if err != nil {
-		klog.Errorf("Failed to unmount dir: %v", err)
-		return err
-	}
-	err = os.Remove(sharedDir)
-	if err != nil {
-		klog.Errorf("Failed to remove dir: %v", err)
-		return err
-	}
-
-	return nil
-}
-
-func getShortSharedDir(uid types.UID, volumeName string) string {
-	return path.Join(util.DefaultHostVhostuserBaseDir, string(uid), volumeName)
 }
 
 func (csh cniServerHandler) createOrUpdateIPCr(podRequest request.CniRequest, subnet, ip, macAddr string) error {
@@ -381,10 +317,8 @@ func (csh cniServerHandler) createOrUpdateIPCr(podRequest request.CniRequest, su
 }
 
 func (csh cniServerHandler) handleDel(req *restful.Request, resp *restful.Response) {
-	podRequest := request.CniRequest{}
-	err := req.ReadEntity(&podRequest)
-
-	if err != nil {
+	var podRequest request.CniRequest
+	if err := req.ReadEntity(&podRequest); err != nil {
 		errMsg := fmt.Errorf("parse del request failed %v", err)
 		klog.Error(errMsg)
 		if err := resp.WriteHeaderAndEntity(http.StatusBadRequest, request.CniResponse{Err: errMsg.Error()}); err != nil {
@@ -392,18 +326,19 @@ func (csh cniServerHandler) handleDel(req *restful.Request, resp *restful.Respon
 		}
 		return
 	}
+
 	pod, err := csh.Controller.podsLister.Pods(podRequest.PodNamespace).Get(podRequest.PodName)
-	if err != nil && !k8serrors.IsNotFound(err) {
+	if err != nil {
+		if k8serrors.IsNotFound(err) {
+			resp.WriteHeader(http.StatusNoContent)
+			return
+		}
+
 		errMsg := fmt.Errorf("parse del request failed %v", err)
 		klog.Error(errMsg)
 		if err := resp.WriteHeaderAndEntity(http.StatusBadRequest, request.CniResponse{Err: errMsg.Error()}); err != nil {
 			klog.Errorf("failed to write response, %v", err)
 		}
-		return
-	}
-
-	if k8serrors.IsNotFound(err) {
-		resp.WriteHeader(http.StatusNoContent)
 		return
 	}
